+++ Sources/KeyboardKit/Extensions/String+Casing.swift
@@ -24,6 +24,27 @@ public extension String {
      Whether or not the string is uppercased.
      */
     var isUppercased: Bool { self == uppercased() && self != lowercased() }
+
+
+    /**
+     Case-adjust the string for the provided `text`.
+
+     This will match capitalization, uppercase and lowercase
+     states in that order, meaning that a single, uppercased
+     character is interpreted as capitalized, not uppercased.
+     */
+    func caseAdjusted(for text: String) -> String {
+        if text.isCapitalized {
+            return self.capitalized
+        }
+        if text == text.uppercased() {
+            return self.uppercased()
+        }
+        if text == text.lowercased() {
+            return self.lowercased()
+        }
+        return self
+    }
 }
 
 public extension Sequence where Iterator.Element == String {