+++ Tests/KeyboardKitTests/Extensions/String+CasingTests.swift
@@ -58,5 +58,32 @@ class String_UppercasedTests: QuickSpec {
                 expect(array.uppercased()).to(equal(expected))
             }
         }
+
+        describe("case adjusted for text") {
+
+            func result(for string: String, text: String) -> String {
+                string.caseAdjusted(for: text)
+            }
+
+            it("is capitalized for single uppercased char") {
+                let result = result(for: "testing this", text: "T")
+                expect(result).to(equal("Testing This"))
+            }
+
+            it("is capitalized for capitalized text") {
+                let result = result(for: "testing this", text: "Test")
+                expect(result).to(equal("Testing This"))
+            }
+
+            it("is uppercased for uppercased text") {
+                let result = result(for: "testing this", text: "TEST")
+                expect(result).to(equal("TESTING THIS"))
+            }
+
+            it("is lowercased for lowercased text") {
+                let result = result(for: "TESTING THIS", text: "test")
+                expect(result).to(equal("testing this"))
+            }
+        }
     }
 }